drop duplicate outputs after a single input
For some reason, the SpeechRecognizer spits out the same callback
repeatedly sometimes.  Now, a set of deltas after an input can only be
filled with unique outputs.  This is perfectly reasonable for the
SpeechRecognizer, and is probably fine for others?  It's at least more
correct than repeated callbacks that should only be seen once.